change table pk as long

+++ src/migration/mysql/scripts/20101129094603_create_table_orders.sql
@@ -1,7 +1,7 @@
 --// create table orders
 -- Migration SQL that makes the change goes here.
 create table orders(
-    id int(11) NOT NULL auto_increment,
+    id bigint(11) NOT NULL auto_increment,
     name varchar(20),
     note varchar(200),
     wf_id varchar(20),
